setup env files
